LES NEGATIFS

+++ instructions.c
@@ -64,10 +64,10 @@ void executer_add(int32_t instruction)
 }
 void executer_addi(int32_t instruction)
 {
-	printf("C'est un ADDI $%d,$%d,%d\n", (instruction & 0x001F0000) >> 16, (instruction & 0x03E00000) >> 21, instruction & 0x0000FFFF);
+	printf("C'est un ADDI $%d,$%d,%d\n", (instruction & 0x001F0000) >> 16, (instruction & 0x03E00000) >> 21, (int16_t)(instruction & 0x0000FFFF));
 	//Integer Overflow ??
 	int32_t temp;
-	temp = lectureRegistre((instruction & 0x001F0000) >> 16) + lectureRegistre((instruction & 0x03E00000) >> 21) + (instruction & 0x0000FFFF);
+	temp = lectureRegistre((instruction & 0x001F0000) >> 16) + lectureRegistre((instruction & 0x03E00000) >> 21) + (int16_t)(instruction & 0x0000FFFF);
 	ecritureRegistre((instruction & 0x001F0000) >> 16,temp);
 }
 void executer_and(int32_t instruction)
@@ -77,27 +77,27 @@ void executer_and(int32_t instruction)
 }
 void executer_beq(int32_t instruction)
 {
-	printf("C'est un BEQ $%d,$%d,%d\n", (instruction & 0x03E00000) >> 21, (instruction & 0x001F0000) >> 16, instruction & 0x0000FFFF);
+	printf("C'est un BEQ $%d,$%d,%d\n", (instruction & 0x03E00000) >> 21, (instruction & 0x001F0000) >> 16, (int16_t)(instruction & 0x0000FFFF));
 	if (lectureRegistre((instruction & 0x03E00000) >> 21) == lectureRegistre((instruction & 0x001F0000) >> 16))
-		PC += ((instruction & 0x0000FFFF) << 2);
+		PC += ((int16_t)(instruction & 0x0000FFFF) << 2);
 }
 void executer_bgtz(int32_t instruction)
 {
-	printf("C'est un BGTZ $%d,%d\n", (instruction & 0x03E00000) >> 21, (instruction & 0x0000FFFF));
+	printf("C'est un BGTZ $%d,%d\n", (instruction & 0x03E00000) >> 21, (int16_t)(instruction & 0x0000FFFF));
 	if (lectureRegistre((instruction & 0x03E00000) >> 21) > 0)
-		PC += ((instruction & 0x0000FFFF) << 2);
+		PC += ((int16_t)(instruction & 0x0000FFFF) << 2);
 }
 void executer_blez(int32_t instruction)
 {
-	printf("C'est un BLEZ $%d,%d\n", (instruction & 0x03E00000) >> 21, instruction & 0x0000FFFF);
+	printf("C'est un BLEZ $%d,%d\n", (instruction & 0x03E00000) >> 21, (int16_t)(instruction & 0x0000FFFF));
 	if (lectureRegistre((instruction & 0x03E00000) >> 21) <= 0)
-		PC += ((instruction & 0x0000FFFF) << 2);
+		PC += ((int16_t)(instruction & 0x0000FFFF) << 2);
 }
 void executer_bne(int32_t instruction)
 {
-	printf("C'est un BNE $%d,$%d,%d\n", (instruction & 0x03E00000) >> 21, (instruction & 0x001F0000) >> 16, instruction & 0x0000FFFF);
+	printf("C'est un BNE $%d,$%d,%d\n", (instruction & 0x03E00000) >> 21, (instruction & 0x001F0000) >> 16, (int16_t)(instruction & 0x0000FFFF));
 	if (lectureRegistre((instruction & 0x03E00000) >> 21) != lectureRegistre((instruction & 0x001F0000) >> 16))
-		PC += ((instruction & 0x0000FFFF) << 2);
+		PC += ((int16_t)(instruction & 0x0000FFFF) << 2);
 }
 void executer_div(int32_t instruction)
 {
@@ -128,13 +128,13 @@ void executer_jr(int32_t instruction)
 }
 void executer_lui(int32_t instruction)
 {
-	printf("C'est un LUI $%d,%d\n", (instruction & 0x001F0000) >> 16, instruction & 0x0000FFFF);
-	ecritureRegistre((instruction & 0x001F0000) >> 16, (instruction & 0x0000FFFF) << 16);
+	printf("C'est un LUI $%d,%d\n", (instruction & 0x001F0000) >> 16, (int16_t)(instruction & 0x0000FFFF));
+	ecritureRegistre((instruction & 0x001F0000) >> 16, (int16_t)(instruction & 0x0000FFFF) << 16);
 }
 void executer_lw(int32_t instruction)
 {
-	printf("C'est un LW $%d,%d($%d)\n", (instruction & 0x001F0000) >> 16, instruction & 0x0000FFFF, (instruction & 0x03E00000) >> 21);
-	int16_t offset = instruction & 0x0000FFFF;
+	printf("C'est un LW $%d,%d($%d)\n", (instruction & 0x001F0000) >> 16, (int16_t)(instruction & 0x0000FFFF), (instruction & 0x03E00000) >> 21);
+	int16_t offset = (int16_t)(instruction & 0x0000FFFF);
 	int32_t address = lectureRegistre((instruction & 0x03E00000) >> 21) + offset;
 	if ((address & 0x00000003) != 0)
 		printf("Exception, adresse incorrecte\n");
@@ -207,8 +207,8 @@ void executer_sub(int32_t instruction)
 }
 void executer_sw(int32_t instruction)
 {
-	printf("C'est un SW $%d,%d($%d)\n", (instruction & 0x001F0000) >> 16, instruction & 0x0000FFFF, (instruction & 0x03E00000) >> 21);
-	int16_t offset = instruction & 0x0000FFFF;
+	printf("C'est un SW $%d,%d($%d)\n", (instruction & 0x001F0000) >> 16, (int16_t)(instruction & 0x0000FFFF), (instruction & 0x03E00000) >> 21);
+	int16_t offset = (int16_t)(instruction & 0x0000FFFF);
 	int32_t address = lectureRegistre((instruction & 0x03E00000) >> 21) + offset;
 	if ((address & 0x00000003) != 0)
 		printf("Exception, adresse incorrecte\n");
